requirements.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
caer==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
comtypes==1.1.14
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.6
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
Faker==16.8.1
fastjsonschema==2.16.2
Flask==2.2.2
flatbuffers==22.9.24
fonttools==4.37.4
fqdn==1.5.1
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.4
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
MouseInfo==0.1.3
mypy==1.0.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.1
opencv-contrib-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pydot==1.4.2
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
Pyro4==4.82
pyrsistent==0.19.2
PyScreeze==0.1.28
python-dateutil==2.8.2
python-json-logger==2.0.4
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7
pywhatkit==5.4
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
Send2Trash==1.8.0
serpent==1.41
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
